+++ src/main/kotlin/data/repositories/ProjectsRepositoryImpl.kt
@@ -1,5 +1,5 @@
 package data.repositories
 
-import logic.repositoies.ProjectsRepository
+import logic.repositories.ProjectsRepository
 
 class ProjectsRepositoryImpl: ProjectsRepository
