+++ src/main/kotlin/logic/repositories/AuditRepository.kt
@@ -1,4 +1,4 @@
-package logic.repositoies
+package logic.repositories
 
 interface AuditRepository {
 }
