+++ src/main/kotlin/logic/repositories/AuthenticationRepository.kt
@@ -1,4 +1,4 @@
-package logic.repositoies
+package logic.repositories
 
 interface AuthenticationRepository {
