+++ src/main/kotlin/logic/usecases/AuthenticationUseCase.kt
@@ -1,6 +1,6 @@
 package logic.usecases
 
-import logic.repositoies.AuthenticationRepository
+import logic.repositories.AuthenticationRepository
 
 class AuthenticationUseCase(
     private val authenticationRepository: AuthenticationRepository
